some file renamings
